pyproject.toml
[project]
name = "euna-scraper"
package-mode = false
version = "0.1.0"
description = ""
authors = [
    {name = "Mehwish",email = "39596277+mehwishh247@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "pydoll-python (>=2.12.4,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
